Add day 4 in C++.  Allot Point<short>.

+++ 2024/day08.cpp
@@ -36,7 +36,7 @@ bool TEST = false;
 int WIDTH = -1;
 int HEIGHT = -1;
 
-typedef Point<int> point_t;
+typedef Point<short> point_t;
 
 typedef map<char, vector<point_t>> spots_t;
